Logic moved to library modules #21
